Use service in controller and remove serializer
